MitiGAN_single_label/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main.conda
_tflow_select=2.3.0=mkl.conda
absl-py=0.11.0=py37h89c1867_0
astor=0.8.1=py37_0.conda
attrs=20.2.0=pyh9f0ad1d_0
blas=1.0=mkl.conda
brotlipy=0.7.0=py37hb5d75c8_1001
bzip2=1.0.8=h7b6447c_0.conda
c-ares=1.15.0=h7b6447c_1001.conda
ca-certificates=2020.12.8=h06a4308_0.conda
cairo=1.14.12=h8948797_3.conda
certifi=2020.12.5=py37h06a4308_0.conda
cffi=1.14.1=py37he30daa8_0.conda
cftime=1.3.0=py37ha21ca33_0
chardet=3.0.4=py37_1003.conda
cloudpickle=1.5.0=py_0.conda
cryptography=2.9.2=py37h1ba5d50_0.conda
cudatoolkit=10.2.89=hfd86e86_1.conda
curl=7.71.1=hbc83047_1.conda
cxxfilt=0.2.2=py37h3340039_0
cycler=0.10.0=py37_0.conda
cython=0.29.21=py37he6710b0_0.conda
cytoolz=0.10.1=py37h7b6447c_0.conda
dask-core=2.23.0=py_0.conda
dbus=1.13.16=hb2f20db_0.conda
decorator=4.4.2=py_0.conda
expat=2.2.9=he6710b0_2.conda
ffmpeg=4.0=hcdf2ecd_0.conda
fontconfig=2.13.0=h9420a91_0.conda
freeglut=3.0.0=hf484d3e_5.conda
freetype=2.10.2=h5ab3b9f_0.conda
gast=0.2.2=py_0
glib=2.65.0=h3eb4bd4_0.conda
google-pasta=0.2.0=py_0.conda
graphite2=1.3.14=h23475e2_0.conda
grpcio=1.31.0=py37hf8bcb03_0.conda
gst-plugins-base=1.14.0=hbbd80ab_1.conda
gstreamer=1.14.0=hb31296c_0.conda
h5py=2.8.0=py37h989c5e5_3.conda
harfbuzz=1.8.8=hffaf4a1_0.conda
hdf4=4.2.13=hf30be14_1003
hdf5=1.10.2=hba1933b_1.conda
icu=58.2=he6710b0_3.conda
idna=2.10=py_0.conda
imageio=2.9.0=py_0.conda
importlib-metadata=1.7.0=py37hc8dfbb8_0
importlib_metadata=1.7.0=0
iniconfig=1.0.1=pyh9f0ad1d_0
intel-openmp=2020.1=217.conda
jasper=2.0.14=h07fcdf6_1.conda
joblib=0.17.0=py_0
jpeg=9d=h36c2ea0_0
keras-applications=1.0.8=py_1.conda
keras-preprocessing=1.1.0=py_1.conda
kiwisolver=1.2.0=py37hfd86e86_0.conda
krb5=1.18.2=h173b8e3_0.conda
lcms2=2.11=hcbb858e_1
ld_impl_linux-64=2.33.1=h53a641e_7.conda
libcurl=7.71.1=h20c2e04_1.conda
libedit=3.1.20191231=h14c3975_1.conda
libffi=3.3=he6710b0_2.conda
libgcc-ng=9.1.0=hdf63c60_0.conda
libgfortran-ng=7.3.0=hdf63c60_0.conda
libglu=9.0.0=hf484d3e_1.conda
libnetcdf=4.6.1=h628ed10_200
libopencv=3.4.2=hb342d67_1.conda
libopus=1.3.1=h7b6447c_0.conda
libpng=1.6.37=hbc83047_0.conda
libprotobuf=3.13.0=hd408876_0.conda
libsodium=1.0.18=h7b6447c_0.conda
libssh2=1.9.0=h1ba5d50_1.conda
libstdcxx-ng=9.1.0=hdf63c60_0.conda
libtiff=4.1.0=h2733197_1.conda
libuuid=1.0.3=h1bed415_2.conda
libvpx=1.7.0=h439df22_0.conda
libxcb=1.14=h7b6447c_0.conda
libxml2=2.9.10=he19cac6_1.conda
lz4-c=1.9.2=he6710b0_1.conda
markdown=3.2.2=py37_0.conda
matplotlib=3.1.3=py37_0.conda
matplotlib-base=3.1.3=py37hef1b27d_0.conda
meshio=4.3.6=pyhd8ed1ab_0
mkl=2020.1=217.conda
mkl-service=2.3.0=py37he904b0f_0.conda
mkl_fft=1.1.0=py37h23d657b_0.conda
mkl_random=1.1.1=py37h0573a6f_0.conda
more-itertools=8.5.0=py_0
ncurses=6.2=he6710b0_1.conda
netcdf4=1.4.1=py37ha292673_200
ninja=1.10.0=py37hfd86e86_0.conda
numpy=1.19.2=py37h54aff64_0.conda
numpy-base=1.19.2=py37hfa32c7d_0.conda
nvidia-apex=0.1=py37h14a324e_1
olefile=0.46=py37_0.conda
opencv=3.4.2=py37h6fd60c2_1.conda
openssl=1.1.1i=h27cfd23_0.conda
opt_einsum=3.3.0=py_0
packaging=20.4=pyh9f0ad1d_0
pandas=1.0.1=py37h0573a6f_0.conda
pcre=8.44=he6710b0_0.conda
pillow=6.1.0=py37h6b7be26_1
pip=20.2.4=py_0
pixman=0.40.0=h7b6447c_0.conda
pluggy=0.13.1=py37hc8dfbb8_2
plyfile=0.7.2=pyh9f0ad1d_0
protobuf=3.13.0=py37hf484d3e_0.conda
py=1.9.0=pyh9f0ad1d_0
py-opencv=3.4.2=py37hb342d67_1.conda
pycparser=2.20=py_2.conda
pyopenssl=19.1.0=py_1.conda
pyparsing=2.4.7=py_0.conda
pyqt=5.9.2=py37h05f1152_2.conda
pysocks=1.7.1=py37_1.conda
pytest=6.0.1=py37hc8dfbb8_0
python=3.7.7=hcff3b4d_5.conda
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.6.0=py3.7_cuda10.2.89_cudnn7.6.5_0
pytz=2020.1=py_0.conda
pywavelets=1.1.1=py37h7b6447c_0.conda
pyyaml=5.3=py37h7b6447c_0
pyzmq=19.0.1=py37he6710b0_1.conda
qt=5.9.7=h5867ecd_1.conda
readline=8.0=he28a2e2_2
requests=2.24.0=py_0.conda
scikit-learn=0.23.2=py37h0573a6f_0
scipy=1.5.2=py37h0b6359f_0.conda
seaborn=0.11.1=pyhd3eb1b0_0.conda
setuptools=49.6.0=py37_0.conda
sip=4.19.8=py37hf484d3e_0.conda
six=1.15.0=py_0.conda
sqlite=3.33.0=h62c20be_0.conda
tabulate=0.8.7=pyh9f0ad1d_0
tensorboard=1.15.0=py37_0
tensorboardx=2.0=py_0
tensorflow=1.15.0=mkl_py37h28c19af_0.conda
tensorflow-base=1.15.0=mkl_py37he1670d9_0.conda
tensorflow-estimator=1.15.1=pyh2649769_0.conda
termcolor=1.1.0=py37_1.conda
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0.conda
toml=0.10.1=pyh9f0ad1d_0
toolz=0.10.0=py_0.conda
torchfile=0.1.0=py_0
torchvision=0.7.0=py37_cu102
tornado=6.0.4=py37h7b6447c_1.conda
tqdm=4.53.0=pyhd8ed1ab_1
urllib3=1.25.11=py_0
visdom=0.1.8.9=0
websocket-client=0.57.0=py37_1.conda
werkzeug=0.16.1=py_0
wheel=0.34.2=py37_0.conda
wrapt=1.12.1=py37h7b6447c_1.conda
xz=5.2.5=h7b6447c_0.conda
yaml=0.1.7=h96e3832_1
zeromq=4.3.2=he6710b0_2.conda
zipp=3.4.0=py_0
zlib=1.2.11=h7b6447c_3.conda
zstd=1.4.5=h9ceee32_0.conda
